Unions may have empty traces
We don't currently do that, ever, but it is in principle fine, and
should not be ruled out for no apparent reason.

For example, we have discussed maybe using Unions with empty traces to
encode expressions that do not come from the source code, for
example the expressions in the preamble.