Rendering of structuredLang (#7)
* Add dependencies for rendering testing

* Rename `preslang` -> `preslang_old`

Paving the way for the new structLang stuff that is going on over at the
compiler side.

* Basic component with accompanying test for PoC of Item

The implementation of <Item /> is broken, obviously. This is just to see
whether I could get the tests running at all.

* Adds a few more tests, focusing on simple nesting and tuples

* Add more tests, focusing on inner lists a bit

* Update faulty test and add new test

I had mixed up how to handle bare vs. non-bare elements in `args`.

* Fix spacing in the tests

I want to test that the space is there. I hope this way works.

* Relax the proptype-checking for the item component a bit

They will be more strictly checked in the specialized helper functions
that I will create for rendering tuples and constructors, respectively.

* Create the helper functions, with dummy implementations

* Rename `preslang` folder to `structuredlang`

Since that is what it is called in the compiler.

* Moved <Cons /> and <Tuple /> into their own files

* Realized that this test was faulty

* Set up tests for using the helpers from `jest-enzyme`

* Fix stupid typo in test

* Use convenient method from `jest-enzyme` in this failing test

* Damnit, I had forgotten that this structure will actually be nested

* Adds test for checking somewhat proper insertion of spaces between args

* Tests relating to simple constructors are passing now

* Make React shut up about keys

At least for the four tests that are currently passing.

* Throw out <Cons />, intent to start over

Fuck you, React. These fucking key props are making me dumber by the
minute.

* Had forgotten to change the tests so they use jest-enzyme

* Big rethinking of <Item /> (renamed to <StructuredExp />)

1.  Rename <Item /> to <StructuredExp />.
2.  Add a new <List /> component.
3.  Throw out whatever implementation I had of <Cons /> and <Tuple />.
4.  Update the tests to also include wrapping spans for lists.

* Update tests to use `shallow` instead of `mount`

For some reason, this made the first test (that was previously) passing
to fail. Oh well, this tests seem broken anyway, so I will just try to
verify the correct output manually.

* Implement <Item /> so that it "passes" the first four tests

The tests are not "passing", but they return the correct HTML, so eh,
whatever.

* Oops, this test was broken by `toContainReact`

* Implement <Tuple />, tests manually verified to "pass"

* Fix even more broken tests

Damn it, JSX is a blessing and a curse.

* Implement <List />, seems to be passing the tests

* Refactor <List /> and <Tuple /> using new <SurroundAndSeparate />

Behaviour seems to be unchanged.

* Skip all tests for <StructuredExp />

Something is broken about them, it is not our fault, so let's just skip
them for now.

* Replace example JSON with new structuredLang JSON

* Replace <PresLangTrees /> with <StructuredLangTrees /> in container

Currently, <StructuredLangTrees /> is basically a no-op.

* Move the recursive rendering to its own folder

In order to keep the folders a bit more clean and separate unrelated
files from each other (unrelated in some sense of the word).

* Create dummy <pre> wrapper for <StructuredExp />, and use it

* Update import path for `ArrayUtils`

* Oops, had forgotten to use PropTypes for <SurroundAndSeparate />

* Flip the switch! Actually render the trees

This seems to work surprisingly well. The rendering is also fairly fast
for the example JSON -- we will see how well it works for the output
coming from "actual" compilation.

* Remove the old code for rendering presLang.

Oh happy days!

* Fix typo

* Delete `ReactUtils` (and tests) now that they are no longer needed

* Remove some functions from `ArrayUtils` (along with tests)

They were only used by the functions in ReactUtils, but now that
ReactUtils is gone, these are no longer needed.