Basic `onClick` functionality for rendered structuredLang (#8)
* Inject `dispatch` into <StructuredLangTrees />

Note that the actual dispatch function is not injected, yet, but it will
be trivial to substitute later.

* Inject `createOnClick` into <StructuredExpWrapper />

* Inject `createOnClick` into top-level <StructuredExp /> component

We still get warnings, though, since the helper components do not
recieve `createOnClick`, so they cannot pass it to their recursive calls
to <StructuredExp />.

* Inject `createOnClick` into <Item />, <List /> and <Tuple />

We still get boatloads of warnings though, since they have only been
injected into these components, but they are not actually used.

* Forgot about the helper component <SurroundAndSeparate />

* Inject all the `createOnClick`s

We are still not using it anywhere, but we do not get any prop type
failures anymore, so I assume the injections are working.

* Actually create and use onClick functions

Had forgotten about stopping the propagation, but that was easily fixed.

* Update the TRACE_ACTIVATED action

In some other branch, I had changed it to include the `lang` property,
but apparently that branch hasn't been merged.

* Update the shape of the state tree to include lang for highlightedTrace

* Dispatch events when spans are clicked

Woohoo, this seems to work flawlessly! (Here's to hoping I didn't jinx
it by writing that.)

* More gracefully handle clicking on stuff without traces

It it not the most graceful way, but it is better than what was before.